requirements.txt
pandas
google-cloud-storage
tqdm
PyPDF2
python-dotenv
transformers
torch
faiss-cpu
numpy
streamlit